src/products/serializers/product.py
from rest_framework import serializers
from src.products.models.product import Product


class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = '__all__'
