requirements/base.txt
Django==2.2.7
djangorestframework==3.10.3
pylint-django==2.0.13
